requirements.txt
flask
crewai
crewai-tools
boto3
humanize
python-dotenv
langchain
langchain_core
langchain_google_genai
Flask==2.3.3
Flask-Login
Flask-WTF
Flask-Bcrypt
WTForms
Flask-SQLALchemy==2.5.1
bcrypt
SQLALchemy==1.4.39
yfinance